ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
google-genai>=1.0.0
openai>=1.40.0
voyageai>=0.3.0
qdrant-client>=1.11.0
httpx>=0.27.0
pillow>=10.0.0
imagehash>=4.3.0
supabase>=2.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
